add more modules using ES6
