+++ index.html
@@ -22,7 +22,7 @@
     <main>
         <div class="date" id="currdate"></div> 
         <section class="book">        
-        <h2>All Awesome Books List</h2>  
+        <h2>All Awesome Books</h2>  
             <div class="allbooks">        
             </div>
          
@@ -48,9 +48,9 @@
                 </ul>          
         </section>
     </main>
-    <footer>
-            <h4>Copyright</h4>
-             <h3>By Supreeti</h3>
+    <footer class="copyright">
+           <div>Copyright@2023 </div>
+           <div><a target="_blank" href="https://github.com/supreeti">By Supreeti</a></div>  
     </footer>
     <script type="module" src="index.js"></script>   
 </body>
